Multi-page app structure
The app structure was modified to use pages. This was done to be able to add a confirmation page with a link to donate to Achilles, and allows for the addition of other pages in the future to enhance the experience.

The main component in app.tsx was moved to pages/home.tsx, and renamed to Home. app.tsx was replaced with a router component that serves pages to the user. Entering tacobell50k.com will automatically route to the home page.